Fix review issues: dedup, safety, correctness, and performance
- Make Engine::mark_writable public, remove duplication in NVS restore
- Enforce NVS blob size limit (4000B) with early return instead of warn
- Use Release/Acquire ordering on dirty flag for proper happens-before
- Log errors on sensor write failures instead of silently discarding
- Change DHT11 fields from f64 to f32 to avoid software FP emulation
- Improve DHT11 response-wait comments with datasheet timing steps
- Add Preconditions doc section for DHT11 1s minimum read interval
- Add json to esp feature deps to fix feature gate alignment

+++ src/dht11.rs
@@ -11,8 +11,8 @@ use esp_idf_svc::hal::gpio::AnyIOPin;
 /// A successful DHT11 reading.
 #[derive(Debug, Clone)]
 pub struct Dht11Reading {
-    pub temperature: f64,
-    pub humidity: f64,
+    pub temperature: f32,
+    pub humidity: f32,
 }
 
 /// Errors from reading the DHT11 sensor.
@@ -59,7 +59,14 @@ fn wait_for_level(
 /// Read temperature and humidity from a DHT11 sensor.
 ///
 /// The pin must be configured as `InputOutput<OpenDrain>` (open-drain mode).
-/// The caller should wait at least 1 second between reads (DHT11 sampling rate).
+///
+/// # Preconditions
+///
+/// The caller **must** wait at least **1 second** between successive calls.
+/// The DHT11 hardware requires this minimum sampling interval; calling more
+/// frequently will produce timeouts or corrupt readings. There is no runtime
+/// guard — the caller is responsible for enforcing the interval (e.g. via a
+/// `sleep` or timer in the main loop).
 pub fn read_dht11(
     pin: &mut PinDriver<'_, AnyIOPin, InputOutput<OpenDrain>>,
 ) -> Result<Dht11Reading, Dht11Error> {
@@ -68,12 +75,20 @@ pub fn read_dht11(
     Ets::delay_us(20_000);
     pin.set_high().map_err(|_| Dht11Error::NoResponse)?;
 
-    // --- Wait for sensor response ---
-    // Sensor pulls low for ~80 us, then high for ~80 us
-    Ets::delay_us(40); // Wait past host release time
-    wait_for_level(pin, false)?; // Wait for sensor to pull low
-    wait_for_level(pin, true)?;  // Low ~80 us
-    wait_for_level(pin, false)?; // High ~80 us
+    // --- Wait for sensor response (DHT11 datasheet timing) ---
+    // After host releases the bus, the line floats high briefly.
+    // Step 1: 40 μs delay — skip past the host release period while
+    //         the bus is still high (pull-up) before the sensor responds.
+    Ets::delay_us(40);
+    // Step 2: Sensor pulls low for ~80 μs as its "response" signal.
+    //         Wait until the line goes low (sensor has started responding).
+    wait_for_level(pin, false)?;
+    // Step 3: Sensor releases the line — it goes high for ~80 μs as the
+    //         "preparation" period before data transmission begins.
+    wait_for_level(pin, true)?;
+    // Step 4: Sensor pulls low again to start the first data bit's ~50 μs
+    //         low preamble. After this, we enter the 40-bit data loop.
+    wait_for_level(pin, false)?;
 
     // --- Read 40 data bits ---
     let mut data = [0u8; 5];
@@ -100,8 +115,8 @@ pub fn read_dht11(
 
     // DHT11: data[0] = humidity integer, data[1] = humidity decimal (usually 0)
     //        data[2] = temperature integer, data[3] = temperature decimal
-    let humidity = data[0] as f64 + data[1] as f64 * 0.1;
-    let temperature = data[2] as f64 + (data[3] & 0x7F) as f64 * 0.1;
+    let humidity = data[0] as f32 + data[1] as f32 * 0.1;
+    let temperature = data[2] as f32 + (data[3] & 0x7F) as f32 * 0.1;
     // Bit 7 of data[3] indicates negative temperature
     let temperature = if data[3] & 0x80 != 0 {
         -temperature
